+++ src/ServiceManager.php
@@ -1,9 +1,7 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @link      http://github.com/zendframework/zend-servicemanager for the canonical source repository
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -132,6 +130,13 @@ class ServiceManager implements ServiceLocatorInterface
      * @var bool
      */
     protected $configured = false;
+
+    /**
+     * Cached abstract factories from string.
+     *
+     * @var array
+     */
+    private $cachedAbstractFactories = [];
 
     /**
      * Constructor.
@@ -512,11 +517,17 @@ class ServiceManager implements ServiceLocatorInterface
     {
         foreach ($abstractFactories as $abstractFactory) {
             if (is_string($abstractFactory) && class_exists($abstractFactory)) {
-                $abstractFactory = new $abstractFactory();
+                //Cached string
+                if (! isset($this->cachedAbstractFactories[$abstractFactory])) {
+                    $this->cachedAbstractFactories[$abstractFactory] = new $abstractFactory();
+                }
+
+                $abstractFactory = $this->cachedAbstractFactories[$abstractFactory];
             }
 
             if ($abstractFactory instanceof Factory\AbstractFactoryInterface) {
-                $this->abstractFactories[] = $abstractFactory;
+                $abstractFactoryObjHash = spl_object_hash($abstractFactory);
+                $this->abstractFactories[$abstractFactoryObjHash] = $abstractFactory;
                 continue;
             }