+++ test/CommonServiceLocatorBehaviorsTrait.php
@@ -1,9 +1,7 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @link      http://github.com/zendframework/zend-servicemanager for the canonical source repository
+ * @copyright Copyright (c) 2015-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -21,6 +19,7 @@ use Zend\ServiceManager\Factory\FactoryInterface;
 use Zend\ServiceManager\Factory\InvokableFactory;
 use Zend\ServiceManager\Initializer\InitializerInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
+use ZendTest\ServiceManager\TestAsset\CallTimesAbstractFactory;
 use ZendTest\ServiceManager\TestAsset\FailingAbstractFactory;
 use ZendTest\ServiceManager\TestAsset\FailingFactory;
 use ZendTest\ServiceManager\TestAsset\FailingExceptionWithStringAsCodeFactory;
@@ -138,6 +137,42 @@ trait CommonServiceLocatorBehaviorsTrait
         ]);
 
         $serviceManager->get(DateTime::class);
+    }
+
+    public function testAllowsMultipleInstancesOfTheSameAbstractFactory()
+    {
+        CallTimesAbstractFactory::setCallTimes(0);
+
+        $obj1 = new CallTimesAbstractFactory();
+        $obj2 = new CallTimesAbstractFactory();
+
+        $serviceManager = $this->createContainer([
+            'abstract_factories' => [
+                $obj1,
+                $obj2,
+            ]
+        ]);
+        $serviceManager->addAbstractFactory($obj1);
+        $serviceManager->addAbstractFactory($obj2);
+        $serviceManager->has(stdClass::class);
+
+        $this->assertEquals(2, CallTimesAbstractFactory::getCallTimes());
+    }
+
+    public function testWillReUseAnExistingNamedAbstractFactoryInstance()
+    {
+        CallTimesAbstractFactory::setCallTimes(0);
+
+        $serviceManager = $this->createContainer([
+            'abstract_factories' => [
+                CallTimesAbstractFactory::class,
+                CallTimesAbstractFactory::class,
+            ]
+        ]);
+        $serviceManager->addAbstractFactory(CallTimesAbstractFactory::class);
+        $serviceManager->has(stdClass::class);
+
+        $this->assertEquals(1, CallTimesAbstractFactory::getCallTimes());
     }
 
     public function testCanCreateServiceWithAlias()