+++ test/TestAsset/ClassDependingOnAnInterface.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @link      http://github.com/zendframework/zend-servicemanager for the canonical source repository
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace ZendTest\ServiceManager\TestAsset;
+
+use Zend\ServiceManager\Factory\FactoryInterface;
+
+class ClassDependingOnAnInterface
+{
+    public function __construct(FactoryInterface $factory)
+    {
+    }
+}