+++ test/TestAsset/SampleFactory.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @link      http://github.com/zendframework/zend-servicemanager for the canonical source repository
+ * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace ZendTest\ServiceManager\TestAsset;
+
+use Interop\Container\ContainerInterface;
+use Zend\ServiceManager\Factory\FactoryInterface;
+
+class SampleFactory implements FactoryInterface
+{
+
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    {
+        return new InvokableObject();
+    }
+}